Final local changes and conflict resolved

+++ frontend/src/components/Allproblems.jsx
@@ -31,6 +31,8 @@ const filteredProblems = searchTerm.trim() === ""
  const handleDelete=async(problemId)=>{
     try{
       const token = localStorage.getItem("token");
+    if (!window.confirm("Are you sure you want to delete this problem?")) return;
+
       await axios.delete(`${BASE_URL}/problem/${problemId}`, {
         headers:{Authorization:`Bearer ${token}`}
